working on building api to connect to database. server is currently able to run on an instance of docker-compose
