+++ Dockerfile
@@ -15,7 +15,7 @@ WORKDIR /app
 EXPOSE 5000
 
 # Run the server
-CMD flask run app.py 
+CMD python server.py 
 
 # Docker and docker-compose learned here: 
 # https://medium.com/@audretschjames/understanding-docker-as-if-it-were-a-gameboy-96c96392efbf
