--- database.conf
@@ -1,5 +1,0 @@
-POSTGRES_USER=user1
-POSTGRES_PASSWORD=password
-POSTGRES_HOST=http://127.0.0.1
-POSTGRES_PORT=5432
-POSTGRES_DB=nqueens
