--- front-end/src/requirements.txt 
@@ -1,2 +1,0 @@
-Flask==0.12
-requests==2.18
